Refactor router configuration into separate config file
