+++ app/src/sw/sw.ts
@@ -1,9 +1,9 @@
 self.addEventListener('install', () => {
-  // @ts-ignore
+  // @ts-expect-error service worker global skipWaiting not in TS lib
   self.skipWaiting()
 })
 
 self.addEventListener('activate', () => {
-  // @ts-ignore
+  // @ts-expect-error service worker clients.claim not in TS lib
   self.clients.claim()
 })
